requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web interface
streamlit>=1.25.0

# Configuration and utilities
PyYAML>=6.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For advanced features
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
pandas>=2.0.0

# Optional: For model serving
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
